Move files belonging to modules in dkms package
Call update-initramfs for the dkms package. Otherwise the initrd wont contain
the vmxnet module if you installed the dkms package after the tools package.
Move the module scripts to the dmks package.
The modules in the dkms package should work without the tools now

(I am not so sure about udevadm and where it should be placed)